Pyspark groupBy
In this exercise, you'll use the PySpark package to handle a Spark DataFrame. The data is the same as in previous exercises: participants of Olympic events between 1896 and 2016.

The Spark Dataframe, athlete_events_spark is available in your workspace.

The methods you're going to use in this exercise are:

    .printSchema(): helps print the schema of a Spark DataFrame.
    .groupBy(): grouping statement for an aggregation.
    .mean(): take the mean over each group.
    .show(): show the results.

Task 1    Find out the type of athlete_events_spark.
Task 2    Find out the schema of athlete_events_spark.
Task 3    Print out the mean age of the Olympians, grouped by year. Notice that spark has not actually calculated anything yet. You can call this lazy evaluation.
Task 4    Take the previous result and call .show() on the result to calculate the mean age.