Fetch from API
Extract data from an API by sending a request to the API and parsing the response in JSON format by using the requests library to send a request to the Hacker News API.

Hacker News is a social news aggregation website, specifically for articles related to computer science or the tech world in general. Each post on the website has a JSON representation, which you'll see in the response of the request in the exercise.

 
Task 1    Use the correct method of the requests module to get the Hacker News post's JSON object.
Task 2    Print out the response, parsed as a JSON.
Task 3   Assign the "score" key of the post to post_score.