Data extracttion from database
Create and use a function that extracts a database table into a pandas DataFrame object. The tables you'll be extracting are:

    film: the films that are rented out in the DVD store.
    customer: the customers that rented films at the DVD store.

In order to connect to the database, you'll have to use a PostgreSQL connection URI, which looks something like this:

postgresql://[user[:password]@][host][:port][/database]


Task 1 Complete the extract_table_to_pandas() function definition to include the table name in the query.
Task 2 Fill in the connection URI. The host is localhost and port is 5432. The username and password are repl and password, respectively. The database is pagila.
Task 3 Complete the function calls of extract_table_to_pandas() to extract the film and customer tables.