Loading data into PostgreSQL data warehouse
Write out some data to a PostgreSQL data warehouse. That could be useful when you have a result of some transformations, and you want to use it in an application.

For example, the result of a transformation could have added a column with film recommendations, and you want to use them in your online store.
There's a pandas DataFrame called film_pdf in your workspace.