General Notes for the course
These are my Notes for the Streamlined Data Ingestion with pandas. I have compiled from varius sources and written some of them as to make my understanding clear. Please refer it but do not think of them as perfect and error free. I welcome help and suggestion to make this a exhaustive Notes for a bigginer.